+++ codingame/http/async_.py
@@ -27,9 +27,9 @@ class AsyncHTTPClient(BaseHTTPClient):
     async def close(self):
         await self.__session.close()
 
-    async def request(self, service: str, func: str, json: list = []):
+    async def request(self, service: str, func: str, parameters: list = []):
         url = self.BASE + service + "/" + func
-        async with self.__session.post(url, json=json) as response:
+        async with self.__session.post(url, json=parameters) as response:
             data = await response.json()
             try:
                 response.raise_for_status()