+++ tests/sync/test_client.py
@@ -5,6 +5,7 @@ import pytest
 from codingame import exceptions
 from codingame.clash_of_code import ClashOfCode
 from codingame.client import Client
+from codingame.client.sync import SyncClient
 from codingame.codingamer import CodinGamer
 from codingame.leaderboard import (
     ChallengeLeaderboard,
@@ -39,11 +40,30 @@ async def test_client_context_manager_error():
             pass  # pragma: no cover
 
 
-def test_client_login(client: Client, mock_http):
+def test_client_request(client: SyncClient):
+    session = client.request("session", "findSession")
+    assert isinstance(session, dict)
+
+
+@pytest.mark.parametrize(
+    ["service", "func"],
+    [
+        ("", ""),
+        ("session", ""),
+        ("", "findSession"),
+    ],
+)
+def test_client_request_error(
+    client: SyncClient, service: str, func: str
+):
+    with pytest.raises(ValueError):
+        client.request(service, func)
+
+
+def test_client_login(client: SyncClient, mock_http):
     mock_http(client._state.http, "login")
     mock_http(client._state.http, "get_codingamer_from_id")
     mock_http(client._state.http, "get_codingamer_from_handle")
-
     client.login(
         remember_me_cookie=os.environ.get("TEST_LOGIN_REMEMBER_ME_COOKIE"),
     )
@@ -51,7 +71,6 @@ def test_client_login(client: Client, mock_http):
     assert client.codingamer is not None
 
 
-@pytest.mark.skip
 @pytest.mark.parametrize(
     ["email", "password"],
     [
@@ -63,41 +82,7 @@ def test_client_login(client: Client, mock_http):
         (os.environ.get("TEST_LOGIN_EMAIL"), "BadPassword"),
     ],
 )
-def test_client_login_error(
-    client: Client,
-    email: str,
-    password: str,
-    is_mocking: bool,
-    mock_httperror,
-):
-    if is_mocking:
-        if email == "":
-            error = {"id": 332, "message": "Email is required"}
-
-        elif email == "NotAnEmail":
-            error = {"id": 334, "message": "Malformed email"}
-
-        elif email == "nonexistant@example.com":
-            if password == "":
-                error = {"id": 336, "message": "Password is required"}
-
-            elif password == "NonExistant":
-                error = {
-                    "id": 393,
-                    "message": (
-                        "This email address is not linked to a "
-                        "CodinGamer account"
-                    ),
-                }
-
-        elif email == os.environ.get("TEST_LOGIN_EMAIL"):
-            error = {
-                "id": 396,
-                "message": "The password you entered is incorrect.",
-            }
-
-        mock_httperror(client._state.http, "login", error)
-
+def test_client_login_error(client: SyncClient, email: str, password: str):
     with pytest.raises(exceptions.LoginError):
         client.login(email, password)
 
@@ -110,7 +95,7 @@ def test_client_login_error(
         os.environ.get("TEST_CODINGAMER_PUBLIC_HANDLE"),
     ],
 )
-def test_client_get_codingamer(client: Client, codingamer_query, mock_http):
+def test_client_get_codingamer(client: SyncClient, codingamer_query, mock_http):
     mock_http(client._state.http, "search")
     mock_http(client._state.http, "get_codingamer_from_id")
     mock_http(client._state.http, "get_codingamer_from_handle")
@@ -128,7 +113,7 @@ def test_client_get_codingamer(client: Client, codingamer_query, mock_http):
     ],
 )
 def test_client_get_codingamer_error(
-    client: Client, codingamer_query, mock_http, mock_httperror
+    client: SyncClient, codingamer_query, mock_http, mock_httperror
 ):
     mock_http(client._state.http, "search", [])
     mock_httperror(client._state.http, "get_codingamer_from_id", {"id": 404})
@@ -138,14 +123,14 @@ def test_client_get_codingamer_error(
         client.get_codingamer(codingamer_query)
 
 
-def test_client_get_clash_of_code(client: Client):
+def test_client_get_clash_of_code(client: SyncClient):
     clash_of_code = client.get_clash_of_code(
         os.environ.get("TEST_CLASHOFCODE_PUBLIC_HANDLE")
     )
     assert isinstance(clash_of_code, ClashOfCode)
 
 
-def test_client_get_clash_of_code_error(client: Client):
+def test_client_get_clash_of_code_error(client: SyncClient):
     with pytest.raises(ValueError):
         client.get_clash_of_code("0")
 
@@ -153,60 +138,60 @@ def test_client_get_clash_of_code_error(client: Client):
         client.get_clash_of_code("0" * 7 + "a" * 32)
 
 
-def test_client_get_pending_clash_of_code(client: Client):
+def test_client_get_pending_clash_of_code(client: SyncClient):
     clash_of_code = client.get_pending_clash_of_code()
     assert isinstance(clash_of_code, ClashOfCode) or clash_of_code is None
 
 
-def test_client_get_language_ids(client: Client, mock_http):
+def test_client_get_language_ids(client: SyncClient, mock_http):
     mock_http(client._state.http, "get_language_ids")
     language_ids = client.get_language_ids()
     assert isinstance(language_ids, list)
     assert all(isinstance(language_id, str) for language_id in language_ids)
 
 
-def test_client_get_unseen_notifications(auth_client: Client):
+def test_client_get_unseen_notifications(auth_client: SyncClient):
     for notification in auth_client.get_unseen_notifications():
         assert isinstance(notification, Notification)
         assert not notification.seen
         assert not notification.read
 
 
-def test_client_get_unseen_notifications_error(client: Client):
+def test_client_get_unseen_notifications_error(client: SyncClient):
     with pytest.raises(exceptions.LoginRequired):
         next(client.get_unseen_notifications())
 
 
-def test_client_get_unread_notifications(auth_client: Client):
+def test_client_get_unread_notifications(auth_client: SyncClient):
     for notification in auth_client.get_unread_notifications():
         assert isinstance(notification, Notification)
         assert not notification.read
 
 
-def test_client_get_unread_notifications_error(client: Client):
+def test_client_get_unread_notifications_error(client: SyncClient):
     with pytest.raises(exceptions.LoginRequired):
         next(client.get_unread_notifications())
 
 
-def test_client_get_read_notifications(auth_client: Client):
+def test_client_get_read_notifications(auth_client: SyncClient):
     for notification in auth_client.get_read_notifications():
         assert isinstance(notification, Notification)
         assert notification.seen
         assert notification.read
 
 
-def test_client_get_read_notifications_error(client: Client):
+def test_client_get_read_notifications_error(client: SyncClient):
     with pytest.raises(exceptions.LoginRequired):
         next(client.get_read_notifications())
 
 
-def test_client_get_global_leaderboard(client: Client):
+def test_client_get_global_leaderboard(client: SyncClient):
     global_leaderboard = client.get_global_leaderboard()
     assert isinstance(global_leaderboard, GlobalLeaderboard)
     assert isinstance(global_leaderboard.users[0], GlobalRankedCodinGamer)
 
 
-def test_client_get_global_leaderboard_error(client: Client):
+def test_client_get_global_leaderboard_error(client: SyncClient):
     with pytest.raises(ValueError):
         client.get_global_leaderboard(type="NONEXISTENT")
     with pytest.raises(ValueError):
@@ -218,7 +203,9 @@ def test_client_get_global_leaderboard_error(client: Client):
 @pytest.mark.parametrize(
     "challenge_id", ["coders-strike-back", "spring-challenge-2021"]
 )
-def test_client_get_challenge_leaderboard(client: Client, challenge_id: str):
+def test_client_get_challenge_leaderboard(
+    client: SyncClient, challenge_id: str
+):
     challenge_leaderboard = client.get_challenge_leaderboard(challenge_id)
     assert isinstance(challenge_leaderboard, ChallengeLeaderboard)
     assert isinstance(challenge_leaderboard.users[0], ChallengeRankedCodinGamer)
@@ -226,7 +213,7 @@ def test_client_get_challenge_leaderboard(client: Client, challenge_id: str):
         assert isinstance(challenge_leaderboard.leagues[0], League)
 
 
-def test_client_get_challenge_leaderboard_error(client: Client):
+def test_client_get_challenge_leaderboard_error(client: SyncClient):
     with pytest.raises(ValueError):
         client.get_challenge_leaderboard(
             "spring-challenge-2021", group="nonexistent"
@@ -240,7 +227,7 @@ def test_client_get_challenge_leaderboard_error(client: Client):
 
 
 @pytest.mark.parametrize("puzzle_id", ["coders-strike-back", "codingame-optim"])
-def test_client_get_puzzle_leaderboard(client: Client, puzzle_id: str):
+def test_client_get_puzzle_leaderboard(client: SyncClient, puzzle_id: str):
     puzzle_leaderboard = client.get_puzzle_leaderboard(puzzle_id)
     assert isinstance(puzzle_leaderboard, PuzzleLeaderboard)
     assert isinstance(puzzle_leaderboard.users[0], PuzzleRankedCodinGamer)
@@ -258,7 +245,7 @@ def test_client_get_puzzle_leaderboard(client: Client, puzzle_id: str):
         )
 
 
-def test_client_get_puzzle_leaderboard_error(client: Client):
+def test_client_get_puzzle_leaderboard_error(client: SyncClient):
     with pytest.raises(ValueError):
         client.get_puzzle_leaderboard("codingame-optim", group="nonexistent")
     with pytest.raises(exceptions.LoginRequired):